fix: using usleep instead of condition

+++ philo/utils/utils.h
@@ -21,7 +21,7 @@
 int		ft_atoi(const char *nptr);
 int		on_error(int err, t_prog prog, t_app_state state);
 void	clean_up(t_prog prog, t_app_state state);
-int		get_time(void);
+long	get_time(void);
 int		is_arg_valid(char **argv);
 
 #endif
